Bugfixes & Feature Adds
Resolved a couple of bugs with using the progress file. Added feature to ask user whether to use progress file (if present) or not. Added feature to ask user to delete progress file at the end of the scrape. Added feature to ask user how many rows to scrape rather than using a hardcoded value.